+++ src/components/Header/UserItems.tsx
@@ -1,10 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import { isLoginAtom } from 'recoil/atom';
 import styled from 'styled-components';
-import { ReactComponent as PlusIcon } from 'assets/plus-icon.svg';
-import { ReactComponent as BellIcon } from 'assets/bell-icon.svg';
 import { ReactComponent as UserIcon } from 'assets/user-icon.svg';
 import { BREAK_POINT } from 'constants/style';
 import PostOptions from './PostOptions';
@@ -12,8 +10,6 @@ import Notification from './Notification';
 
 const UserItems = () => {
   const [isLogin, setIsLogin] = useRecoilState(isLoginAtom);
-  const [isPostOptionsOpen, setIsPostOptionsOpen] = useState(false);
-  const [isNotificationOpen, setIsNotificationOpen] = useState(false);
 
   const logout = () => {
     sessionStorage.clear();
@@ -22,12 +18,7 @@ const UserItems = () => {
 
   return (
     <Container>
-      {isLogin && (
-        <BellBtn onClick={() => setIsNotificationOpen(true)} onBlur={() => setIsNotificationOpen(false)}>
-          <BellIcon />
-        </BellBtn>
-      )}
-      {isNotificationOpen && <Notification />}
+      <Notification />
       {isLogin && <StyledLink to="/my">마이페이지</StyledLink>}
       <StyledUserIcon login={isLogin} />
       {isLogin ? (
@@ -37,11 +28,7 @@ const UserItems = () => {
       ) : (
         <StyledLink to="/login">로그인 / 회원가입</StyledLink>
       )}
-      <PostBtn onClick={() => setIsPostOptionsOpen(true)} onBlur={() => setIsPostOptionsOpen(false)}>
-        <StyledPlusIcon />
-        글쓰기
-      </PostBtn>
-      {isPostOptionsOpen && <PostOptions />}
+      <PostOptions />
     </Container>
   );
 };
@@ -58,17 +45,6 @@ const Container = styled.div`
   }
 `;
 
-const BellBtn = styled.button`
-  width: 24px;
-  height: 24px;
-  cursor: pointer;
-
-  @media (min-width: ${BREAK_POINT.MOBILE}) {
-    width: 26px;
-    height: 26px;
-  }
-`;
-
 const StyledLink = styled(Link)`
   font-size: 16px;
   font-weight: 400;
@@ -76,31 +52,6 @@ const StyledLink = styled(Link)`
 
   @media (min-width: ${BREAK_POINT.TABLET}) {
     display: block;
-  }
-`;
-
-const PostBtn = styled.button`
-  background-color: #fceec0;
-  border-radius: 10px;
-  padding: 8px 12px;
-  display: none;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  color: #000;
-  font-size: 18px;
-  font-weight: 500;
-
-  @media (min-width: ${BREAK_POINT.TABLET}) {
-    display: flex;
-  }
-`;
-
-const StyledPlusIcon = styled(PlusIcon)`
-  width: 16px;
-  height: 16px;
-  & path {
-    stroke: #000;
   }
 `;
